+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - 添加 ErrorType 用于表示该语法错误的类型；
 - LSP 的 textDocument/publishDiagnostics 添加了 DiagnosticTag 的支持；
 - 添加 core.Ranger 接口的定义；
+- 添加了对 textDocument/references 的支持；
 
 ## [v7.1.0]